+++ app.js
@@ -1,37 +1,37 @@
-const [command, key, value] = process.argv.slice(2);
-const { readPasswords, writePasswords } = require("./lib/passwords");
+const [command] = process.argv.slice(2);
+const { get, set, unset } = require("./lib/commands");
 
-function get() {
-  console.log("Called GET", key);
-  try {
-    const passwords = readPasswords();
-    console.log(key, passwords[key]);
-  } catch (error) {
-    console.error(error);
-  }
-}
+// function get() {
+//   console.log("Called GET", key);
+//   try {
+//     const passwords = readPasswords();
+//     console.log(key, passwords[key]);
+//   } catch (error) {
+//     console.error(error);
+//   }
+// }
 
-function set() {
-  console.log("Called SET", key, value);
-  try {
-    const passwords = readPasswords();
-    passwords[key] = value;
-    writePasswords(passwords);
-  } catch (error) {
-    console.error(error);
-  }
-}
+// function set() {
+//   console.log("Called SET", key, value);
+//   try {
+//     const passwords = readPasswords();
+//     passwords[key] = value;
+//     writePasswords(passwords);
+//   } catch (error) {
+//     console.error(error);
+//   }
+// }
 
-function unset() {
-  console.log("Called UNSET", key, value);
-  try {
-    const passwords = readPasswords();
-    delete passwords[key];
-    writePasswords(passwords);
-  } catch (error) {
-    console.error(error);
-  }
-}
+// function unset() {
+//   console.log("Called UNSET", key, value);
+//   try {
+//     const passwords = readPasswords();
+//     delete passwords[key];
+//     writePasswords(passwords);
+//   } catch (error) {
+//     console.error(error);
+//   }
+// }
 
 if (command === "get") {
   get();
